+++ src/components/AnimatedParagraph.tsx
