+++ src/components/Contributors.tsx
@@ -1,10 +1,42 @@
 "use client";
 
+import { useEffect, useState } from "react";
 import Image from "next/image";
 import contributorsData from "@/data/contributors.json";
 
-export default function Contributors() {
-  const contributors = contributorsData.contributors;
+// The roster is refreshed by the @elixpoo bot (see
+// .github/workflows/update-contributors.yml) and served raw from main.
+// We read it live so new contributors show up without a redeploy, and fall
+// back to the bundled copy for first paint / offline / rate-limited reads.
+const RAW_ROSTER_URL =
+  "https://raw.githubusercontent.com/elixpo/elixpo/main/src/data/contributors.json";
+
+export function Contributors() {
+  const [contributors, setContributors] = useState<string[]>(
+    contributorsData.contributors,
+  );
+  const [subheadline, setSubheadline] = useState<string>(
+    contributorsData.subheadline,
+  );
+
+  useEffect(() => {
+    let active = true;
+    fetch(RAW_ROSTER_URL, { cache: "no-store" })
+      .then((res) => (res.ok ? res.json() : null))
+      .then((data) => {
+        if (!active || !data || !Array.isArray(data.contributors) || !data.contributors.length) {
+          return;
+        }
+        setContributors(data.contributors);
+        if (typeof data.subheadline === "string") setSubheadline(data.subheadline);
+      })
+      .catch(() => {
+        /* keep the bundled fallback */
+      });
+    return () => {
+      active = false;
+    };
+  }, []);
 
   // Split contributors into orbital rings
   const orbit1 = contributors.slice(0, 6);
@@ -12,14 +44,23 @@ export default function Contributors() {
   const orbit3 = contributors.slice(15);
 
   return (
-    <section className="py-24 md:py-32 overflow-hidden">
-      <div className="max-w-7xl mx-auto px-6">
+    <section
+      id="contributors"
+      className="bg-black text-[#E1E0CC] py-24 md:py-32 px-4 sm:px-6 lg:px-8 relative overflow-hidden select-none"
+    >
+      {/* Ambient background glow */}
+      <div className="absolute top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 w-[50vw] h-[50vw] bg-primary/[0.03] rounded-full blur-[140px] pointer-events-none" />
+
+      <div className="max-w-7xl mx-auto relative z-10">
         <div className="text-center mb-16">
-          <h2 className="text-4xl md:text-6xl font-bold tracking-tight mb-4">
-            {contributorsData.headline}
+          <span className="text-[10px] uppercase tracking-widest text-[#DEDBC8] font-medium font-mono block mb-3">
+            The People Behind Elixpo
+          </span>
+          <h2 className="text-3xl sm:text-4xl md:text-5xl font-light tracking-tight text-white mb-4">
+            Our <span className="italic font-serif text-primary">Contributors</span>
           </h2>
-          <p className="text-muted max-w-xl mx-auto text-base">
-            {contributorsData.subheadline}
+          <p className="text-[#DEDBC8]/60 max-w-xl mx-auto text-xs sm:text-sm font-mono leading-relaxed">
+            {subheadline}
           </p>
         </div>
 
@@ -32,13 +73,13 @@ export default function Contributors() {
               alt="Circuit-Overtime"
               width={80}
               height={80}
-              className="w-16 h-16 md:w-20 md:h-20 rounded-full border-4 border-accent shadow-xl shadow-accent/30"
+              className="w-16 h-16 md:w-20 md:h-20 rounded-full border-4 border-primary shadow-xl shadow-primary/30"
               unoptimized
             />
           </div>
 
           {/* Orbit 1 - innermost */}
-          <div className="absolute inset-[25%] rounded-full border border-border/40">
+          <div className="absolute inset-[25%] rounded-full border border-white/10">
             <div className="relative w-full h-full animate-orbit-slow">
               {orbit1.map((username, i) => {
                 const angle = (i / orbit1.length) * 360;
@@ -60,7 +101,7 @@ export default function Contributors() {
                       alt={username}
                       width={36}
                       height={36}
-                      className="w-8 h-8 md:w-9 md:h-9 rounded-full border-2 border-white shadow-md"
+                      className="w-8 h-8 md:w-9 md:h-9 rounded-full border-2 border-[#DEDBC8]/40 shadow-md"
                       unoptimized
                     />
                   </div>
@@ -70,7 +111,7 @@ export default function Contributors() {
           </div>
 
           {/* Orbit 2 - middle */}
-          <div className="absolute inset-[12%] rounded-full border border-border/30">
+          <div className="absolute inset-[12%] rounded-full border border-white/[0.07]">
             <div className="relative w-full h-full animate-orbit-medium">
               {orbit2.map((username, i) => {
                 const angle = (i / orbit2.length) * 360;
@@ -92,7 +133,7 @@ export default function Contributors() {
                       alt={username}
                       width={32}
                       height={32}
-                      className="w-7 h-7 md:w-8 md:h-8 rounded-full border-2 border-white shadow-md"
+                      className="w-7 h-7 md:w-8 md:h-8 rounded-full border-2 border-[#DEDBC8]/30 shadow-md"
                       unoptimized
                     />
                   </div>
@@ -102,7 +143,7 @@ export default function Contributors() {
           </div>
 
           {/* Orbit 3 - outermost */}
-          <div className="absolute inset-0 rounded-full border border-border/20">
+          <div className="absolute inset-0 rounded-full border border-white/[0.05]">
             <div className="relative w-full h-full animate-orbit-outer">
               {orbit3.map((username, i) => {
                 const angle = (i / orbit3.length) * 360;
@@ -124,7 +165,7 @@ export default function Contributors() {
                       alt={username}
                       width={28}
                       height={28}
-                      className="w-6 h-6 md:w-7 md:h-7 rounded-full border-2 border-white shadow-sm"
+                      className="w-6 h-6 md:w-7 md:h-7 rounded-full border-2 border-[#DEDBC8]/20 shadow-sm"
                       unoptimized
                     />
                   </div>
@@ -135,8 +176,8 @@ export default function Contributors() {
         </div>
 
         {/* Contributor count */}
-        <p className="text-center text-sm text-muted mt-8">
-          {contributors.length}+ contributors and growing
+        <p className="text-center text-xs text-[#DEDBC8]/65 font-mono mt-8">
+          35+ contributors and growing
         </p>
       </div>
     </section>
