+++ src/components/Footer.tsx
@@ -1,146 +1,164 @@
-import Link from "next/link";
-import footerData from "@/data/footer.json";
+"use client";
 
-function GitHubIcon() {
+import { useState } from "react";
+import { Copy, Check, Github, MessagesSquare, Heart } from "lucide-react";
+import { ELIXPO_LINKS } from "@/lib/elixpo-links";
+
+const ecosystem = [
+  { label: "Elixpo Art", href: ELIXPO_LINKS.generate },
+  { label: "Elixpo Blogs", href: ELIXPO_LINKS.blog },
+  { label: "LixSketch", href: ELIXPO_LINKS.sketch },
+  { label: "Elixpo Chat", href: ELIXPO_LINKS.chat },
+  { label: "Elixpo Search", href: ELIXPO_LINKS.search },
+  { label: "Elixpo Accounts", href: ELIXPO_LINKS.accounts },
+];
+
+const community = [
+  { label: "GitHub Org", href: ELIXPO_LINKS.github },
+  { label: "Chapter Monorepo", href: ELIXPO_LINKS.githubChapter },
+  { label: "GitHub Discussions", href: ELIXPO_LINKS.discussions },
+  { label: "GitHub Sponsors", href: ELIXPO_LINKS.sponsors },
+];
+
+export function Footer() {
+  const [copied, setCopied] = useState(false);
+
+  const copyEmail = async () => {
+    try {
+      await navigator.clipboard.writeText(ELIXPO_LINKS.email);
+      setCopied(true);
+      setTimeout(() => setCopied(false), 1800);
+    } catch {
+      /* clipboard unavailable */
+    }
+  };
+
   return (
-    <svg className="w-5 h-5" fill="currentColor" viewBox="0 0 24 24">
-      <path d="M12 2C6.477 2 2 6.484 2 12.017c0 4.425 2.865 8.18 6.839 9.504.5.092.682-.217.682-.483 0-.237-.008-.868-.013-1.703-2.782.605-3.369-1.343-3.369-1.343-.454-1.158-1.11-1.466-1.11-1.466-.908-.62.069-.608.069-.608 1.003.07 1.531 1.032 1.531 1.032.892 1.53 2.341 1.088 2.91.832.092-.647.35-1.088.636-1.338-2.22-.253-4.555-1.113-4.555-4.951 0-1.093.39-1.988 1.029-2.688-.103-.253-.446-1.272.098-2.65 0 0 .84-.27 2.75 1.026A9.564 9.564 0 0112 6.844c.85.004 1.705.115 2.504.337 1.909-1.296 2.747-1.027 2.747-1.027.546 1.379.202 2.398.1 2.651.64.7 1.028 1.595 1.028 2.688 0 3.848-2.339 4.695-4.566 4.943.359.309.678.92.678 1.855 0 1.338-.012 2.419-.012 2.747 0 .268.18.58.688.482A10.019 10.019 0 0022 12.017C22 6.484 17.522 2 12 2z" />
-    </svg>
-  );
-}
+    <footer className="bg-black border-t border-white/10 px-4 sm:px-6 lg:px-8 pt-16 pb-10 relative overflow-hidden">
+      {/* Ambient brand glow */}
+      <div className="absolute -bottom-24 left-1/2 -translate-x-1/2 w-[60vw] h-[240px] bg-primary/[0.05] rounded-full blur-[120px] pointer-events-none" />
 
-function DiscussionsIcon() {
-  return (
-    <svg className="w-5 h-5" fill="currentColor" viewBox="0 0 16 16" aria-hidden="true">
-      <path d="M1.75 1h8.5c.966 0 1.75.784 1.75 1.75v5.5A1.75 1.75 0 0 1 10.25 10H7.061l-2.574 2.573A1.458 1.458 0 0 1 2 11.543V10h-.25A1.75 1.75 0 0 1 0 8.25v-5.5C0 1.784.784 1 1.75 1ZM1.5 2.75v5.5c0 .138.112.25.25.25h1a.75.75 0 0 1 .75.75v2.19l2.72-2.72a.749.749 0 0 1 .53-.22h3.5a.25.25 0 0 0 .25-.25v-5.5a.25.25 0 0 0-.25-.25h-8.5a.25.25 0 0 0-.25.25Zm13 2a.25.25 0 0 0-.25-.25h-.5a.75.75 0 0 1 0-1.5h.5c.966 0 1.75.784 1.75 1.75v5.5A1.75 1.75 0 0 1 14.25 12H14v1.543a1.458 1.458 0 0 1-2.487 1.03L9.22 12.28a.749.749 0 0 1 1.06-1.06l2.22 2.22v-2.19a.75.75 0 0 1 .75-.75h1a.25.25 0 0 0 .25-.25Z" />
-    </svg>
-  );
-}
+      <div className="max-w-7xl mx-auto relative z-10">
+        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-10 lg:gap-8 mb-14">
 
-function TwitterIcon() {
-  return (
-    <svg className="w-5 h-5" fill="currentColor" viewBox="0 0 24 24">
-      <path d="M18.244 2.25h3.308l-7.227 8.26 8.502 11.24H16.17l-5.214-6.817L4.99 21.75H1.68l7.73-8.835L1.254 2.25H8.08l4.713 6.231zm-1.161 17.52h1.833L7.084 4.126H5.117z" />
-    </svg>
-  );
-}
-
-function MailIcon() {
-  return (
-    <svg className="w-5 h-5" fill="none" stroke="currentColor" strokeWidth={1.5} viewBox="0 0 24 24">
-      <path strokeLinecap="round" strokeLinejoin="round" d="M21.75 6.75v10.5a2.25 2.25 0 01-2.25 2.25h-15a2.25 2.25 0 01-2.25-2.25V6.75m19.5 0A2.25 2.25 0 0019.5 4.5h-15a2.25 2.25 0 00-2.25 2.25m19.5 0v.243a2.25 2.25 0 01-1.07 1.916l-7.5 4.615a2.25 2.25 0 01-2.36 0L3.32 8.91a2.25 2.25 0 01-1.07-1.916V6.75" />
-    </svg>
-  );
-}
-
-const socialIcons: Record<string, () => React.JSX.Element> = {
-  github: GitHubIcon,
-  discussions: DiscussionsIcon,
-  twitter: TwitterIcon,
-};
-
-function FooterLink({ href, children }: { href: string; children: React.ReactNode }) {
-  const isExternal = href.startsWith("http");
-  if (isExternal) {
-    return (
-      <a
-        href={href}
-        target="_blank"
-        rel="noopener noreferrer"
-        className="text-sm text-gray-400 hover:text-white transition-colors"
-      >
-        {children}
-      </a>
-    );
-  }
-  return (
-    <Link href={href} className="text-sm text-gray-400 hover:text-white transition-colors">
-      {children}
-    </Link>
-  );
-}
-
-export default function Footer() {
-  return (
-    <footer className="bg-foreground text-white">
-      <div className="max-w-7xl mx-auto px-6 py-16">
-        <div className="grid grid-cols-1 md:grid-cols-4 gap-12">
-          {/* Brand + Socials + Contact */}
-          <div className="md:col-span-2 space-y-6">
-            <div>
-              <h3 className="text-2xl font-bold mb-2">{footerData.brand.name}</h3>
-              <p className="text-sm text-gray-400 max-w-sm">{footerData.brand.tagline}</p>
+          {/* Brand block */}
+          <div className="lg:col-span-1">
+            <h3 className="text-2xl font-serif italic text-[#E1E0CC] mb-3">Elixpo Chapter</h3>
+            <p className="text-sm text-[#DEDBC8]/65 leading-relaxed max-w-xs mb-5">
+              An open-source ecosystem of interconnected AI and developer tools, built by a
+              global community of {`35+`} contributors. Free and copyleft, forever.
+            </p>
+            <div className="flex items-center gap-3">
+              <a
+                href={ELIXPO_LINKS.github}
+                target="_blank"
+                rel="noopener noreferrer"
+                aria-label="GitHub"
+                className="p-2 rounded-lg border border-white/10 text-[#DEDBC8]/70 hover:text-white hover:border-white/30 transition-colors"
+              >
+                <Github size={16} />
+              </a>
+              <a
+                href={ELIXPO_LINKS.discussions}
+                target="_blank"
+                rel="noopener noreferrer"
+                aria-label="GitHub Discussions"
+                className="p-2 rounded-lg border border-white/10 text-[#DEDBC8]/70 hover:text-white hover:border-white/30 transition-colors"
+              >
+                <MessagesSquare size={16} />
+              </a>
+              <a
+                href={ELIXPO_LINKS.sponsors}
+                target="_blank"
+                rel="noopener noreferrer"
+                aria-label="Sponsor"
+                className="p-2 rounded-lg border border-white/10 text-[#DEDBC8]/70 hover:text-primary hover:border-primary/30 transition-colors"
+              >
+                <Heart size={16} />
+              </a>
             </div>
-
-            {/* Contact email */}
-            <a
-              href={`mailto:${footerData.contact.email}`}
-              className="inline-flex items-center gap-2 text-sm text-gray-400 hover:text-white transition-colors group"
-            >
-              <MailIcon />
-              <span className="group-hover:underline">{footerData.contact.email}</span>
-            </a>
-
-            {/* Socials */}
-            <div className="flex gap-3">
-              {footerData.socials.map((social) => {
-                const Icon = socialIcons[social.platform];
-                return (
-                  <a
-                    key={social.platform}
-                    href={social.href}
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    className="w-10 h-10 rounded-full bg-white/10 flex items-center justify-center text-gray-400 hover:bg-accent hover:text-white transition-all"
-                    aria-label={social.platform}
-                  >
-                    {Icon && <Icon />}
-                  </a>
-                );
-              })}
-            </div>
-
-            {/* View Projects button */}
-            <Link
-              href="/projects"
-              className="inline-flex items-center gap-2 rounded-full bg-accent px-6 py-2.5 text-sm font-medium text-white hover:bg-accent-light transition-colors"
-            >
-              View Projects
-              <svg className="w-4 h-4" fill="none" stroke="currentColor" strokeWidth={2} viewBox="0 0 24 24">
-                <path strokeLinecap="round" strokeLinejoin="round" d="M13.5 4.5L21 12m0 0l-7.5 7.5M21 12H3" />
-              </svg>
-            </Link>
           </div>
 
-          {/* Link Columns */}
-          {footerData.columns.map((column) => (
-            <div key={column.title}>
-              <h4 className="text-sm font-semibold uppercase tracking-wider mb-4">{column.title}</h4>
-              <ul className="space-y-3">
-                {column.links.map((link) => (
-                  <li key={link.href}>
-                    <FooterLink href={link.href}>{link.label}</FooterLink>
-                  </li>
-                ))}
-              </ul>
-            </div>
-          ))}
+          {/* Ecosystem links */}
+          <div>
+            <h4 className="text-[10px] uppercase tracking-widest font-mono text-[#DEDBC8]/50 mb-4">Ecosystem</h4>
+            <ul className="space-y-2.5">
+              {ecosystem.map((item) => (
+                <li key={item.label}>
+                  <a
+                    href={item.href}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                    className="text-sm text-[#DEDBC8]/75 hover:text-white transition-colors"
+                  >
+                    {item.label}
+                  </a>
+                </li>
+              ))}
+            </ul>
+          </div>
+
+          {/* Community links */}
+          <div>
+            <h4 className="text-[10px] uppercase tracking-widest font-mono text-[#DEDBC8]/50 mb-4">Community</h4>
+            <ul className="space-y-2.5">
+              {community.map((item) => (
+                <li key={item.label}>
+                  <a
+                    href={item.href}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                    className="text-sm text-[#DEDBC8]/75 hover:text-white transition-colors"
+                  >
+                    {item.label}
+                  </a>
+                </li>
+              ))}
+            </ul>
+          </div>
+
+          {/* Developer contact */}
+          <div>
+            <h4 className="text-[10px] uppercase tracking-widest font-mono text-[#DEDBC8]/50 mb-4">Get in touch</h4>
+            <p className="text-sm text-[#DEDBC8]/65 mb-3 leading-relaxed">
+              Questions, partnerships, or just want to say hi? Reach the developers directly.
+            </p>
+            <button
+              onClick={copyEmail}
+              className="group inline-flex items-center gap-2 px-4 py-2.5 rounded-full border border-white/10 bg-white/[0.04] hover:border-primary/30 transition-colors"
+              aria-label="Copy email address"
+            >
+              <span className="text-sm font-mono text-[#DEDBC8] group-hover:text-white transition-colors">
+                {ELIXPO_LINKS.email}
+              </span>
+              {copied ? (
+                <Check size={14} className="text-green-400" />
+              ) : (
+                <Copy size={14} className="text-[#DEDBC8]/60 group-hover:text-primary transition-colors" />
+              )}
+            </button>
+            {copied && (
+              <span className="block text-[10px] font-mono text-green-400/80 mt-2">Copied to clipboard!</span>
+            )}
+          </div>
         </div>
 
-        <div className="mt-14 pt-8 border-t border-gray-800 flex flex-col sm:flex-row items-center justify-between gap-4 text-sm text-gray-500">
-          <p>&copy; {footerData.copyright}</p>
-          <p>
-            Built with &hearts; by the{" "}
-            <a
-              href="https://github.com/elixpo/elixpo_chapter"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="text-gray-400 hover:text-white transition-colors"
-            >
-              Elixpo community
+        {/* Bottom bar */}
+        <div className="border-t border-white/10 pt-8 flex flex-col md:flex-row justify-between items-center gap-4 text-xs">
+          <span className="text-[#DEDBC8]/55 font-mono">
+            © 2026 Elixpo Chapter · GNU GPL-3.0 · Built in the open
+          </span>
+          <div className="flex flex-wrap justify-center gap-x-6 gap-y-2">
+            <a href={ELIXPO_LINKS.generate} target="_blank" rel="noopener noreferrer" className="text-[#DEDBC8]/60 hover:text-white transition-colors">
+              elixpo.com
             </a>
-          </p>
-        </div>
+            <a href={ELIXPO_LINKS.blog} target="_blank" rel="noopener noreferrer" className="text-[#DEDBC8]/60 hover:text-white transition-colors">
+              Blog
+            </a>
+            <a href={ELIXPO_LINKS.githubChapter} target="_blank" rel="noopener noreferrer" className="text-[#DEDBC8]/60 hover:text-white transition-colors">
+              GitHub
+            </a>
+          </div        </div>
       </div>
     </footer>
   );
