+++ src/components/NewsletterSection.tsx
