+++ src/components/NominationSection.tsx
