+++ src/components/PackagesSection.tsx
