+++ src/components/SmoothScroll.tsx
