+++ src/components/legacy/Architecture.tsx
