+++ src/components/legacy/Hero.tsx
